Move CoordsProxy tests to their own file.
